% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{competitors}
\alias{competitors}
\title{Competidores de Lucahura Retail Group / Competitors}
\format{
Un data frame / a data frame con 3 filas y 3 columnas:
\describe{
\item{region}{Region donde opera el competidor / region where the
competitor operates.}
\item{competitor}{Nombre ficticio del competidor / fictional
competitor name.}
\item{opening_date}{Fecha de apertura / opening date.}
}
}
\source{
Datos simulados creados para el paquete datamysteries / simulated
data created for the datamysteries package.
}
\usage{
competitors
}
\description{
Apertura de competidores por region para Lucahura Retail Group.
}
\details{
Competitor openings by region for Lucahura Retail Group.
}
\keyword{datasets}
